[FIRParser/ExportVerilog] Rework the change for Issue #2076
We want ExportVerilog to have predictable behavior no matter what
the input IR is.  The fix to that issue added some special behavior
for 0 values, but the actual problem is all the way up in the parser
StringRef::getAsInteger parser used by the firrtl frontend, which puts
zeros into an arbitrary 64-bit value.  Change that in the .fir parser
to properly confine integer constants to 32-bits when they fit.

This eliminates the firrtl specific logic from the verilog printer.

+++ test/Conversion/ExportVerilog/hw-dialect.mlir
@@ -713,13 +713,13 @@ hw.module @Chi() -> (Chi_output : i0) {
    // CHECK-NEXT: ) bar ();
 
    hw.instance "bar" @Bar1360<
-     WIDTH0: i64 = 0, WIDTH1: i4 = 4, WIDTH2: i40 = 6812312123, WIDTH3: si4 = -1,
+     WIDTH0: i32 = 0, WIDTH1: i4 = 4, WIDTH2: i40 = 6812312123, WIDTH3: si4 = -1,
      WIDTH4: si68 = -88888888888888888, Wtricky: i40 = 4294967295
    >() -> ()
    hw.output
  }
  hw.module.extern @Bar1360<
-     WIDTH0: i64, WIDTH1: i4, WIDTH2: i40, WIDTH3: si4, WIDTH4: si68, Wtricky: i40
+     WIDTH0: i32, WIDTH1: i4, WIDTH2: i40, WIDTH3: si4, WIDTH4: si68, Wtricky: i40
    >() attributes {verilogName = "RealBar"}
 
 // CHECK-LABEL: module Issue1563(
